Fix the URLs generated by docs.help_icon on the progress_effects page.
Without the `with context` the generated URL is relative to the octoprint server, as opposed to being on cp2004.gitbook.io/ws281x-led-status/.

It seems this was a mistake when this change was originally made. Additionally it seems overview.jinga2 has a `with context` where it doesn't need to be.